+++ packages/lumx-react/src/components/button/Button.tsx
@@ -2,9 +2,9 @@ import isEmpty from 'lodash/isEmpty';
 
 import { Emphasis, Icon, Size, Theme, Text, ThemeProvider } from '@lumx/react';
 import { isComponent } from '@lumx/react/utils/type';
-import { getBasicClass } from '@lumx/core/js/utils/_internal/className';
 import type { LumxClassName } from '@lumx/core/js/types';
 import { classNames } from '@lumx/core/js/utils';
+import { modifier } from '@lumx/core/js/utils/classNames/bem/modifier';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
@@ -70,8 +70,10 @@ export const Button = forwardRef<ButtonProps, HTMLButtonElement | HTMLAnchorElem
 
     const buttonClassName = classNames.join(
         className,
-        getBasicClass({ prefix: CLASSNAME, type: 'hasLeftIcon', value: !isEmpty(leftIcon) }),
-        getBasicClass({ prefix: CLASSNAME, type: 'hasRightIcon', value: !isEmpty(rightIcon) }),
+        modifier(CLASSNAME, {
+            'has-left-icon': Boolean(leftIcon),
+            'has-right-icon': Boolean(rightIcon),
+        }),
     );
 
     return (